# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/coupon_codes:
    get:
      tags:
        - CouponCodes
      summary: List CouponCodes
      description: Get a paginated list of CouponCodes with optional filtering and sorting
      operationId: listCouponCodes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponCodeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CouponCodes
      summary: Create CouponCode
      description: Create a new CouponCode
      operationId: createCouponCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCouponCodeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponCode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/coupon_redemptions:
    get:
      tags:
        - CouponRedemptions
      summary: List CouponRedemptions
      description: Get a paginated list of CouponRedemptions with optional filtering and sorting
      operationId: listCouponRedemptions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponRedemptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CouponRedemptions
      summary: Create CouponRedemption
      description: Create a new CouponRedemption
      operationId: createCouponRedemption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCouponRedemptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponRedemption'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    CouponCode:
      type: object
      required:
        - id
        - company_id
        - code
        - pricing_rule_id
        - used_count
        - valid_from
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        pricing_rule_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 280
        max_use:
          nullable: true
          type: integer
          format: int32
        used_count:
          type: integer
          format: int32
        valid_from:
          type: string
          format: date-time
        valid_upto:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    CreateCouponCodeRequest:
      type: object
      required:
        - company_id
        - code
        - pricing_rule_id
        - used_count
        - valid_from
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        pricing_rule_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 280
        max_use:
          nullable: true
          type: integer
          format: int32
        used_count:
          type: integer
          format: int32
        valid_from:
          type: string
          format: date-time
        valid_upto:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    CouponCodeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CouponCode'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CouponRedemption:
      type: object
      required:
        - id
        - company_id
        - coupon_id
        - pricing_rule_id
        - source_type
        - source_id
        - redeemed_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        coupon_id:
          type: string
          format: uuid
        pricing_rule_id:
          type: string
          format: uuid
        source_type:
          type: string
          maxLength: 40
        source_id:
          type: string
          format: uuid
        redeemed_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateCouponRedemptionRequest:
      type: object
      required:
        - company_id
        - coupon_id
        - pricing_rule_id
        - source_type
        - source_id
        - redeemed_at
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        coupon_id:
          type: string
          format: uuid
        pricing_rule_id:
          type: string
          format: uuid
        source_type:
          type: string
          maxLength: 40
        source_id:
          type: string
          format: uuid
        redeemed_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CouponRedemptionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CouponRedemption'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []